Tests/LibWeb/Text/expected/HTML/form-control-type-getters.txt

fieldset.type: fieldset
output.type: output
textarea.type: textarea
select.type: select-one
select.type after setting multiple: select-multiple
